fixe async wrapper, add logout

+++ src/server/actions/deleteTransaction.ts
@@ -1,10 +1,12 @@
 'use server';
 
+import { getAuth } from '@/server/getAuth';
 import { prisma } from '@/server/prisma';
 import { withAuth } from '@/server/withAuth';
 
 const _deleteTransaction = async (id: string) => {
-	await prisma.transaction.delete({ where: { id } });
+	const user = await getAuth();
+	await prisma.transaction.delete({ where: { id, userId: user.id } });
 
 	return id;
 };
